human refactor pass
