+++ nodl_schema/nodl_schema/__init__.py
@@ -13,6 +13,7 @@ from nodl_schema.composition import (
 )
 from nodl_schema.loader import dump_nodl, load_nodl, load_nodl_with_doc_tree, parse_nodl, resolve_document
 from nodl_schema.local_resolver import LocalResolver
+from nodl_schema.rewrite import rewrite_references
 from nodl_schema.validation import load_schema, validate
 
 register_resolver(AmentIndexResolver())
@@ -31,6 +32,7 @@ __all__ = [
     'register_resolver',
     'resolve_document',
     'resolver_registered',
+    'rewrite_references',
     'unregister_resolver',
     'validate',
 ]
