+++ nodl_schema/package.xml
@@ -12,6 +12,7 @@
   <exec_depend>ament_index_python</exec_depend>
   <exec_depend>python3-jsonschema</exec_depend>
   <exec_depend>python3-pydantic</exec_depend>
+  <exec_depend>python3-ruamel.yaml</exec_depend>
   <exec_depend>python3-yaml</exec_depend>
 
   <test_depend>python3-pytest</test_depend>
